+++ src/bin/pass-it-on-server.rs
@@ -1,11 +1,9 @@
 use clap::Parser;
-use tracing::{error, info};
-use pass_it_on::Error;
-use pass_it_on::ServerConfiguration;
-use pass_it_on::{start_server, verify_matrix_devices};
+use pass_it_on::{Error, ServerConfiguration, start_server, verify_matrix_devices};
 use std::path::PathBuf;
 use std::process::ExitCode;
 use tracing::level_filters::LevelFilter;
+use tracing::{error, info};
 
 #[derive(Parser, Debug)]
 #[clap(name = "pass-it-on-server", author, version, about = "Pass-it-on server binary", long_about = None)]