+++ src/interfaces/pipe.rs
@@ -161,7 +161,7 @@ impl Interface for PipeInterface {
 
     #[cfg(not(feature = "pipe-server"))]
     async fn receive(&self, _interface_tx: Sender<String>, _shutdown: watch::Receiver<bool>) -> Result<(), Error> {
-        Err(Error::DisabledInterfaceFeature("pipe-server".to_string()))
+        Err(Error::disabled_interface_feature("pipe-server".to_string()))
     }
 
     #[cfg(feature = "pipe-client")]
@@ -189,7 +189,7 @@ impl Interface for PipeInterface {
         _interface_rx: broadcast::Receiver<Notification>,
         _shutdown: watch::Receiver<bool>,
     ) -> Result<(), Error> {
-        Err(Error::DisabledInterfaceFeature("pipe-client".to_string()))
+        Err(Error::disabled_interface_feature("pipe-client".to_string()))
     }
 }